The the goal of this project is to predict the housing prices of a town or a suburb based on the features of the locality provided to us. In the process, we need to identify the most important features affecting the price of the house. We employed techniques of data preprocessing and built a linear regression model that predicts the prices for the unseen data.
The final model is generalized and perfectly predicts prices with a 100% r-squared. Our model can very accuratly predict the housing prices in Boston and would be a usefull tool in the real estate, banking, and insurance industries.

From EDA and our model we where able to extract that value in Boston houses is primarily measured by:

Areas with low crime rates
Being on the bounds of the Charles River
Older and more indust rial neighboorhoods
Having more rooms
Located near more urban areas